week1/temp/fastapi_streamlit-main/requirements.txt
fastapi==0.103.1
uvicorn
python-multipart
tensorflow
pillow